optimized the structure
